+++ orafce.h
@@ -55,6 +55,6 @@ extern Oid	equality_oper_funcid(Oid argtype);
 
 extern STRING_PTR_FIELD_TYPE ora_days[];
 
-extern int ora_seq_search(const char *name, STRING_PTR_FIELD_TYPE array[], int max);
+extern int ora_seq_search(const char *name, STRING_PTR_FIELD_TYPE array[], size_t max);
 
 #endif
